requirements.txt
flask
opencv-python
numpy
pyserial